Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,7 +2,7 @@
 void main(){
     int a,b,c;
     scanf("%d %d %d",&a,&b,&c);
-    int d;
+    float d;
     d=a+b+c;
     printf("%.2f", d/3);
